Revert "Resolve #331 - moving disconnected champion to the fountain"
This reverts commit 0a45a6cf213b1128e3abf575e45211ecb390116e.

+++ GameServerCore/IGame.cs
@@ -1,6 +1,5 @@
 ﻿using GameServerCore.Maps;
 using GameServerCore.Domain;
-using System.Collections.Generic;
 
 namespace GameServerCore
 {
@@ -14,6 +13,6 @@ namespace GameServerCore
 
         bool SetToExit { get; set; }
 
-        bool HandleDisconnect(ulong userId);
+        bool HandleDisconnect(int userId);
     }
 }
